// client/src/store/action/type/action.ts
export enum CREATE_TYPE_ACTION {
    CREATE_TYPE_PENDING = "CREATE_TYPE_PENDING",
    CREATE_TYPE_SUCCESS = "CREATE_TYPE_SUCCESS",
    CREATE_TYPE_FAILURE = "CREATE_TYPE_FAILURE",
}

export enum GET_ALL_TYPE_ACTION {
    GET_ALL_TYPE_PENDING = "GET_ALL_TYPE_PENDING",
    GET_ALL_TYPE_SUCCESS = "GET_ALL_TYPE_SUCCESS",
    GET_ALL_TYPE_FAILURE = "GET_ALL_TYPE_FAILURE",
}

export enum UPDATE_TYPE_ACTION {
    UPDATE_TYPE_PENDING = "UPDATE_TYPE_PENDING",
    UPDATE_TYPE_SUCCESS = "UPDATE_TYPE_SUCCESS",
    UPDATE_TYPE_FAILURE = "UPDATE_TYPE_FAILURE",
}

export enum DELETE_TYPE_ACTION {
    DELETE_TYPE_PENDING = "DELETE_TYPE_PENDING",
    DELETE_TYPE_SUCCESS = "DELETE_TYPE_SUCCESS",
    DELETE_TYPE_FAILURE = "DELETE_TYPE_FAILURE",
}